Trait sc_piece_cache::PieceCache

pub trait PieceCache: Clone {
    fn add_pieces(
        &self,
        first_piece_index: PieceIndex,
        pieces: &FlatPieces
    ) -> Result<(), Box<dyn Error>>;
    fn get_piece(
        &self,
        piece_index: PieceIndex
    ) -> Result<Option<Piece>, Box<dyn Error>>;
}

Caching layer for pieces produced during archiving to make them available for some time after they were produced.

Required Methods

fn add_pieces(
&self,
first_piece_index: PieceIndex,
pieces: &FlatPieces
) -> Result<(), Box<dyn Error>>

Add pieces to cache

fn get_piece(
&self,
piece_index: PieceIndex
) -> Result<Option<Piece>, Box<dyn Error>>

Get piece from cache